+++ Solution/API/Models/Canton.cs
@@ -14,8 +14,8 @@ namespace API.Models
             Cliente = new HashSet<Cliente>();
         }
 
-        public short CodigoCanton { get; set; }
-        public short CodigoProvincia { get; set; }
+        public int CodigoCanton { get; set; }
+        public int CodigoProvincia { get; set; }
         public string NombreCanton { get; set; }
 
         public virtual Provincia CodigoProvinciaNavigation { get; set; }
