+++ Solution/P.API/Models/Canton.cs
@@ -12,8 +12,8 @@ namespace P.API.Models
            
         }
 
-        public short CodigoCanton { get; set; }
-        public short CodigoProvincia { get; set; }
+        public int CodigoCanton { get; set; }
+        public int CodigoProvincia { get; set; }
         public string NombreCanton { get; set; }
 
         public virtual Provincia CodigoProvinciaNavigation { get; set; }
