+++ Solution/P.DAL.DO/Objects/Cliente.cs
@@ -17,7 +17,7 @@ namespace P.DAL.DO.Objects
         public string SeguApellido { get; set; }
         public string Correo { get; set; }
         public string Direccion { get; set; }
-        public short CodigoCanton { get; set; }
+        public int CodigoCanton { get; set; }
 
         public virtual Canton CodigoCantonNavigation { get; set; }
         public virtual ICollection<ListCal> ListCal { get; set; }
