+++ Solution/P.DAL.DO/Objects/Provincia.cs
@@ -11,7 +11,7 @@ namespace P.DAL.DO.Objects
             Canton = new HashSet<Canton>();
         }
 
-        public short CodigoProvincia { get; set; }
+        public int CodigoProvincia { get; set; }
         public string NombreProvincia { get; set; }
 
         public virtual ICollection<Canton> Canton { get; set; }
